moved new action config github repo to client github folder and trigger action from client
