--- packages/plugin-config-github-info/.npmignore
@@ -1,6 +1,0 @@
-*
-
-!dist/**
-!package.json
-!readme.md
-!tsup.config.ts
