--- packages/plugin-config-github-info/package.json
@@ -1,27 +1,0 @@
-{
-    "name": "@elizaos/plugin-config-github-info",
-    "version": "0.1.0",
-    "main": "dist/index.js",
-    "type": "module",
-    "types": "dist/index.d.ts",
-    "dependencies": {
-        "@elizaos/core": "workspace:*",
-        "@octokit/rest": "20.1.1",
-        "@octokit/types": "12.6.0",
-        "glob": "10.4.5",
-        "simple-git": "3.27.0"
-    },
-    "devDependencies": {
-        "@types/glob": "8.1.0",
-        "eslint": "9.13.0",
-        "eslint-config-prettier": "9.1.0",
-        "eslint-plugin-prettier": "5.2.1",
-        "eslint-plugin-vitest": "0.5.4",
-        "tsup": "8.3.5"
-    },
-    "scripts": {
-        "build": "tsup --format esm --dts",
-        "dev": "tsup --watch",
-        "lint": "eslint . --fix"
-    }
-}
