--- packages/plugin-config-github-info/readme.md
@@ -1,3 +1,0 @@
-# Retrieve GitHub Info Plugin
-
-This plugin retrieves information from GitHub repositories.
