--- packages/plugin-config-github-info/src/utils.ts
@@ -1,5 +1,0 @@
-import path from "path";
-
-export function getRepoPath(owner: string, repo: string) {
-    return path.join("/tmp", "elizaos-repos", owner, repo);
-}
