--- packages/plugin-config-github-info/tsconfig.json
@@ -1,13 +1,0 @@
-{
-    "extends": "../core/tsconfig.json",
-    "compilerOptions": {
-        "outDir": "dist",
-        "rootDir": "src",
-        "types": [
-            "node"
-        ]
-    },
-    "include": [
-        "src/**/*.ts",
-    ]
-}
